Merge of the PreciseBitmapMaterial with the BitmapMaterial. A new parameter is expected from now in BitmapMaterial to set the precision of the perspective correction.
Performance shall not be affected a lot, give me some feedbacks if it is not the case.

From now, BitmapMaterial accepts the LightAttributes as a Flat shading. It is still basic and not amazingly fast, but this is a start.

New demo has been added in order to test the lightening.