Attempt to make the Interation with Flex components. Addition of a FlexInteractionDemo into examples. This example does not work yet, and is something to look into for those who can.
MovieMaterial accepts a Sprite instead of a MovieClip. This makes the class more flexible in my opinion, since UIComponents are extending Sprite and not MovieClip.

Let me know if somebody is able to find out why this is not working with Flex components.

Polygon > modificatin of the swapDepth implementation. Now the normal vector is reversed.
Matrix4 > USE_FAST_MATH is now disabled by default.