Fix on the interaction thing to broadcast only one event, and when there's several,only  the polygon closest to the camera is broadcasted.
removed the neighboors computations from the shape object and added it to the outline attribute.
Now something is missing, the fact to state that this informatino has been computed already. In case several attributes want to do the same, whihc would be unnecessary.

The material manager implementation has changed a bit, and shal work as expected now. (outline atribute is using it)