ok, I could not really get clear win here, but it fixes minor sorting glitch with clipping at the very least :)
to sum up for Thomas, former dictionary version should win in theory because it is accessed less times from camera loop, but then in practice some extra time is wasted for "new Dictionary" call, and with clipping it is not useful at all, because we have to pass cvertices to camera that partially contain same vertices that we would later pass in dictionary (and clipping process is too separate to change that quickly). my "new" logic is actually your old logic from few revisions back + timer cache which turns out not so useful in shape3d as it was in camera, again, because amount of math per vertex we save is small compared to other heavy stuff we do (clipping/etc).

I don't feel attached to changes in this revision (except sorting glitch fix), so feel free to roll back, if in doubt :)